Change n_step_gru converter to use new function patching mechanism
