+++ onnx_chainer/functions/__init__.py
@@ -106,3 +106,5 @@ from onnx_chainer.functions.pooling import convert_MaxPooling2D  # NOQA
 from onnx_chainer.functions.pooling import convert_MaxPoolingND  # NOQA
 from onnx_chainer.functions.pooling import convert_ROIPooling2D  # NOQA
 from onnx_chainer.functions.pooling import convert_Unpooling2D  # NOQA
+
+from onnx_chainer.functions.rnn import convert_n_step_gru  # NOQA
